# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ctf-blockchain-infra
  version: "1.0.0"
  description: |
    HTTP surface of the ctf-blockchain-infra backend. The public
    endpoints (`/api/{config,status,flag,sign,launch,kill,reset,instance,rpc,health}`)
    are unauthenticated and per-IP rate-limited. The admin surface
    (`/api/admin/*`) requires a bearer token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://ctf.example.com
    description: Default deployment URL pattern
  - url: http://127.0.0.1:8787
    description: Local dev backend
...
security: []
paths:
...
  /api/admin/instances:
    get:
      tags: [admin]
      summary: List active private-anvil instances.
      security: [{ AdminBearer: [] }]
      responses:
        "200": { description: OK }
        "401": { description: unauthorized }
  /api/admin/instances/kill-all:
    post:
      tags: [admin]
      summary: Drain every running instance.
      security: [{ AdminBearer: [] }]
      responses:
        "200": { description: OK }
  /api/admin/instances/{challengeId}/{player}/kill:
    post:
      tags: [admin]
      summary: Force-kill one player's instance.
      security: [{ AdminBearer: [] }]
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string }
        - in: path
          name: player
          required: true
          schema: { type: string, format: address }
      responses:
        "200": { description: OK }
  /api/admin/solves:
    get:
      tags: [admin]
      summary: Dump the in-memory solve tracker.
      security: [{ AdminBearer: [] }]
      responses:
        "200": { description: OK }
  /api/admin/challenges:
    get:
      tags: [admin]
      summary: Per-challenge summary (mode, target, signer status).
      security: [{ AdminBearer: [] }]
      responses:
        "200": { description: OK }
  /api/admin/manifest/reload:
    post:
      tags: [admin]
      summary: Re-read challenges.json from disk and reconcile.
      security: [{ AdminBearer: [] }]
      responses:
        "200":
          description: Reload result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reloaded: { type: boolean }
                  added:    { type: array, items: { type: string } }
                  removed:  { type: array, items: { type: string } }
                  total:    { type: integer }
        "500":
          description: Validation failed; previous manifest still active.
  /api/admin/webhook/test:
    post:
      tags: [admin]
      summary: Emit a synthetic webhook event.
      security: [{ AdminBearer: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                event:     { type: string }
                challenge: { type: string }
                player:    { type: string, format: address }
                solved:    { type: boolean }
                previous:
                  oneOf:
                    - { type: boolean }
                    - { type: "null" }
      responses:
        "200": { description: webhook sent }
        "400": { description: webhook not configured }
components:
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
...